pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solo-company-os"
version = "0.1.0"
description = "A runnable learning project: deploy a one-person agent company powered by Agent Skills (SKILL.md)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["agents", "agent-skills", "llm", "workflow", "education"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "PyYAML>=6.0.1",
  "pydantic>=2.7.4",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "ruff>=0.5.5",
]

[project.scripts]
solo-company = "solo_company_os.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
norecursedirs = ["runs", ".venv", "venv"]
